Fixed some compile warnings. Removed old CQ3 cpp file. Changed free and total memory values to kB values (2^10 bytes) to avoid integer wraps on large systems. Replaced memory getter under Linux with POSIX compliant version which also works for SUNs. Moved Linux CPU clock getter to OSX part because it only works for BSD systems, Linux seems to require the /proc/cpuinfo parsing - still to be implemented. Fixed the clean targets of all examples to ignore errors and added a build-all script for all Makefile using systems. Added some more cases for the new matrix 4 constructors, not sure what the rest should do. Removed some Windows stuff from examples.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@387 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CIrrDeviceLinux.cpp
@@ -656,8 +656,8 @@ bool CIrrDeviceLinux::run()
 			{
 			case ConfigureNotify:
 				// check for changed window size
-				if ((event.xconfigure.width != Width) ||
-					(event.xconfigure.height != Height))
+				if ((event.xconfigure.width != (int) Width) ||
+					(event.xconfigure.height != (int) Height))
 				{
 					Width = event.xconfigure.width;
 					Height = event.xconfigure.height;
